+++ tests/Controller/ScenarioScheduler/ScenarioSchedulerControllerTest.php
@@ -29,10 +29,12 @@ final class ScenarioSchedulerControllerTest extends TestCase
         [$controller, $history] = $this->createControllerAndHistory([$response]);
         $returnedMappedResponse = $controller->listScenarioSchedulers(...$methodArguments);
 
+        Assert::same($response->getStatusCode(), $returnedMappedResponse->getStatusCode());
         Assert::equal($mappedResponseBody, $returnedMappedResponse->getBody());
         Assert::same($httpMethod, $history[0]['request']->getMethod());
         Assert::same($requestedUrl, (string) $history[0]['request']->getUri());
         Assert::same($response, $history[0]['response']);
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
     }
 
     /**
@@ -43,10 +45,12 @@ final class ScenarioSchedulerControllerTest extends TestCase
         [$controller, $history] = $this->createControllerAndHistory([$response]);
         $returnedMappedResponse = $controller->getScenarioScheduler(...$methodArguments);
 
+        Assert::same($response->getStatusCode(), $returnedMappedResponse->getStatusCode());
         Assert::equal($mappedResponseBody, $returnedMappedResponse->getBody());
         Assert::same($httpMethod, $history[0]['request']->getMethod());
         Assert::same($requestedUrl, (string) $history[0]['request']->getUri());
         Assert::same($response, $history[0]['response']);
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
     }
 
     /**
@@ -65,6 +69,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
         [$controller, $history] = $this->createControllerAndHistory([$response]);
         $returnedMappedResponse = $controller->createScenarioScheduler($methodArgument);
 
+        Assert::same($response->getStatusCode(), $returnedMappedResponse->getStatusCode());
         Assert::equal($mappedResponseBody, $returnedMappedResponse->getBody());
         Assert::same($httpMethod, $history[0]['request']->getMethod());
         Assert::same($requestedUrl, (string) $history[0]['request']->getUri());
@@ -75,6 +80,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
 
         Assert::same(json_decode($requestBody, true), json_decode((string) $history[0]['request']->getBody(), true));
         Assert::equal($response, $history[0]['response']);
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
         Assert::same($expectedEtag, $returnedMappedResponse->getEtag());
     }
 
@@ -94,6 +100,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
         [$controller, $history] = $this->createControllerAndHistory([$response]);
         $returnedMappedResponse = $controller->updateScenarioScheduler(...$methodArguments);
 
+        Assert::same($response->getStatusCode(), $returnedMappedResponse->getStatusCode());
         Assert::equal($mappedResponseBody, $returnedMappedResponse->getBody());
         Assert::same($httpMethod, $history[0]['request']->getMethod());
         Assert::same($requestedUrl, (string) $history[0]['request']->getUri());
@@ -104,6 +111,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
 
         Assert::same(json_decode($requestBody, true), json_decode((string) $history[0]['request']->getBody(), true));
         Assert::equal($response, $history[0]['response']);
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
         Assert::same($expectedEtag, $returnedMappedResponse->getEtag());
     }
 
@@ -122,6 +130,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
         [$controller, $history] = $this->createControllerAndHistory([$response]);
         $returnedMappedResponse = $controller->validateScenarioScheduler($methodArgument);
 
+        Assert::same($response->getStatusCode(), $returnedMappedResponse->getStatusCode());
         Assert::equal($mappedResponseBody, $returnedMappedResponse->getBody());
         Assert::same($httpMethod, $history[0]['request']->getMethod());
         Assert::same($requestedUrl, (string) $history[0]['request']->getUri());
@@ -132,6 +141,7 @@ final class ScenarioSchedulerControllerTest extends TestCase
 
         Assert::same(json_decode($requestBody, true), json_decode((string) $history[0]['request']->getBody(), true));
         Assert::equal($response, $history[0]['response']);
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
     }
 
     public function testDeleteScenarioScheduler(): void
@@ -140,10 +150,12 @@ final class ScenarioSchedulerControllerTest extends TestCase
             new Response(204, [], null),
         ]);
 
-        $controller->deleteScenarioScheduler('b5f052d8-f285-49ac-9030-bd50ec73393c');
+        $returnedMappedResponse = $controller->deleteScenarioScheduler('b5f052d8-f285-49ac-9030-bd50ec73393c');
 
+        Assert::same(204, $returnedMappedResponse->getStatusCode());
         Assert::same('DELETE', $history[0]['request']->getMethod());
         Assert::same('https://www.crawler.com/api/scenario-schedulers/b5f052d8-f285-49ac-9030-bd50ec73393c', (string) $history[0]['request']->getUri());
+        Assert::same($history[0]['response'], $returnedMappedResponse->getResponse());
     }
 
     public function dataProviderListScenarioSchedulers(): array